Creating metrics from log messages.
This change adds the ability of scanning log messages and incrementing counters
every time a log message matches a given regular expression.

The CLI has a new subcommand `scan` to manage scans, similar to the existing
subcommand for collections.

The store package has been updated with the CRUD operations to manage scans.