feat: reload certificates when creating a new connection
Added a new switch flag when starting the server to enable reloading
certificates, and defaulted to `false` for backward compatibility.

Introduced two new exposed functions to the database factory instead of
changing the original functions signature for both ease of use and not
breaking a lot of places in the codebase.

The reload function will be triggered (if enabled) on every new connection;

Depending on the lifetime of the connection, when it ends, creating a
new connection will always reload the certificates from the path.

The `Reset` function on `sighup` will close all the connections but not
close the pool, allowing new connections to be established.

Signed-off-by: Ihab Soliman <ihab.soliman@form3.tech>