cmd env: adding an env package to aid in testing.
This change adds an Env object that can be injected into the various commands.
The Env can be configured with an in memory store for testing purposes.

This changes also adds a function to inject arbitrary errors in the in memory store
and fixes the getters to avoid a panic on a nil result.