histograms: refactoring histograms rewrite
This change moves the logic that reads prometheus from a source,
applies the histogram translators and rewrites the metric to
a writer in its own `metric` package.

The source now supports a file URL in addition to a HTTP URL to
aid in testing.

A new utility function to load histograms has been added to reduce
code duplication.